automoc: Renaming this to automoc4. Additionally adding a patch to deal with some
kde apps that have a problem with our current version of boost. The instigator of this
was kdepim-4.9.0 and its kleopatra failing on the make. I could have just diddled with
the BUILD of kdepim to fix this but putting it here solves that and any other kde modules
that might have a similar problem with boost.

See http://comments.gmane.org/gmane.comp.kde.devel.pim.bugs/69067