applications: bump to 17.12.0. Move kwave into applications. Many changes to DEPENDS all around.
Yes, I know the kde-application profiles is YUGE. Please bear with me on this, it's not as bad as
it seems. For now with the constant changes to dependencies the KDE folks are doing and their
march to the new frameworks and all that, they are still working on eliminating their monolithic
packages of old.

Right now, if you lin frameworks kde-applications plasma you will get a functional KDE without
qt4/kde4 of old. There may still be a few modules asking for those ancient things, if so, just
say no.