pyside6: A rework of patches. More importantly reworked BUILD. Even
with patches the pathing for with typesystems and glue are incorrect pointing to /usr/typesystems,
etc when they live in /usr/share/pyside6. The prior BUILD twiddle gets overridden with;
PySide6Config.cpython-312-x86_64-linux-gnu.cmake. So a sedit after install of that file pointing
to their correct locations corrects those modules optionally depending on pyside6.